Centrality

This function takes a connectomic edgelist and creates a source and a sink mode- and cell- organized dot plot. The y-axis is the discrete variable 'mode', and the x-axis is the sum of the weights of all edges for each mode made by each cell. Points are organized by cell type, with the size of the point correlating to the Kleinberg hub score (for source graph) and Kleinberg authority score (for sink graph). Network filtration is performed prior to network centrality calculations.

Centrality(
  connectome,
  cols.use = NULL,
  weight.attribute = "weight_sc",
  min.z = NULL,
  normalize = T,
  group.by = "mode",
  ...
)

Arguments

connectome
A connectomic edgelist

cols.use
Desired colors for cell types, alphabetized. Defaults to standard ggplot colors.

weight.attribute
Column to use to define edgeweights for network analysis. 'weight_sc' or 'weight_norm'. Defaults to 'weight_sc'

min.z
Minimum z-score for ligand and receptor.

normalize
Default TRUE. Scales each mode to have equivalent x-axes.

group.by
Default 'mode'. Determines how to subdivide network for centrality analysis. Accepts 'mode','ligand','receptor', 'gene', or 'mechanism'.

...
Arguments passed to FilterConnectome